Updated some columns to reflect pattypan import change
To resolve a conflict in image filenames, I'm switching the input pattypan-generated spreadsheet, to one based on an intermediate fileset, from the /processed folder. Some changes to columns are consequently required. I have also noticed the license column seems to be missing but the permissions column is the one used in the documentation, so have switched to that.